cat_net_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
python-multipart==0.0.20